More extensive refactoring: add SplitBlockViewModel and get rid of dependence on static in Block to cach the select Html code.
