+++ Glyssen/Glyssen.csproj
@@ -459,6 +459,7 @@
     </Compile>
     <Compile Include="Dialogs\RolesForVoiceActorsSaveAsDialog.cs" />
     <Compile Include="Dialogs\SelectBundleForProjectDlg.cs" />
+    <Compile Include="Dialogs\SplitBlockViewModel.cs" />
     <Compile Include="Dialogs\UnappliedSplitsViewModel.cs" />
     <Compile Include="Dialogs\UnappliedSplitsDlg.cs">
       <SubType>Form</SubType>
