+++ GlyssenEngine/Utilities/IFontInfo.cs
@@ -1,0 +1,9 @@
+﻿namespace GlyssenEngine.Utilities
+{
+	public interface IFontInfo
+	{
+		bool RightToLeftScript { get; }
+		string FontFamily { get; }
+		int Size { get; }
+	}
+}
